Remove dup table, get ready for mapping table populating.
This is a pretty big change, probably the last of the data model changes before
merging back into master. Duplicates are expressed with a nullable column,
"dup_id", where if it's not null, it's the ID of the original entry, and the
jot_id of the duplicate is a v4 UUID. Non-duplicate jot_ids are v5 UUIDs under
the NAMESPACE_JOT namespace, generated with the jot's serialized content.

This means that the jot_id is guaranteed to be unique, so it's cool to use as a
primary key. It's also cool, then, to use it as part of a v5 UUID data input,
which is what we're doing for mapping_ids.